server/cement_agent/pyproject.toml
[project]
name = "cement_agent"
version = "0.0.1"
description = "Starter template for making a new agent LangGraph."
authors = []
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.38.0",
    "langchain-google-genai>=2.1.12",
    "langchain-mcp-adapters>=0.1.10",
    "langgraph>=0.2.6",
    "postgres-mcp>=0.3.0",
    "python-dotenv>=1.0.1",
]

[tool.setuptools]
packages = ["langgraph.templates.agent", "agent"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"
